Rework explorer routing
