--- apps/explorer/src/app/App.module.css
@@ -1,20 +1,0 @@
-.app {
-    @apply w-full;
-
-    background: linear-gradient(
-        to bottom,
-        #2a4362 50%,
-        #2a4362 0%,
-        #767a81 0%,
-        #767a81 50%
-    );
-}
-
-main {
-    @apply relative z-10 min-h-[77vh] rounded-[20px]
-    bg-offwhite py-2 shadow-2xl;
-}
-
-.suicontainer {
-    @apply mx-auto max-w-[1440px] p-5;
-}
