+++ apps/explorer/src/index.tsx
@@ -1,92 +1,30 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-    WalletStandardAdapterProvider,
-    UnsafeBurnerWalletAdapter,
-} from '@mysten/wallet-adapter-all-wallets';
-import { WalletProvider, type WalletProviderProps } from '@mysten/wallet-kit';
-import * as Sentry from '@sentry/react';
-import { BrowserTracing } from '@sentry/tracing';
+import '@fontsource/inter/variable.css';
+import '@fontsource/red-hat-mono/variable.css';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter as Router } from 'react-router-dom';
+import { RouterProvider } from 'react-router-dom';
 
-import App from './app/App';
-import { featuresPromise, growthbook } from './utils/growthbook';
+import { router } from './pages';
+import { loadFeatures } from './utils/growthbook';
 import { plausible } from './utils/plausible';
+import './utils/sentry';
 import { reportWebVitals } from './utils/vitals';
 
 import './index.css';
+
+// Start loading features as early as we can:
+loadFeatures();
 
 // NOTE: The plausible tracker ensures it doesn't run on localhost, so we don't
 // need to gate this call.
 plausible.enableAutoPageviews();
 
-if (import.meta.env.PROD) {
-    Sentry.init({
-        dsn: 'https://e4251274d1b141d7ba272103fa0f8d83@o1314142.ingest.sentry.io/6564988',
-        environment: import.meta.env.VITE_VERCEL_ENV,
-        integrations: [new BrowserTracing()],
-        // NOTE: Even though this is set to 1, we actually will properly sample the event in `beforeSendTransaction`.
-        // We don't do sampling here or in `tracesSampler` because those can't be async, so we can't wait for
-        // the features from growthbook to load before applying sampling.
-        tracesSampleRate: 1,
-        async beforeSendTransaction(event) {
-            await featuresPromise;
-            const sampleRate = growthbook.getFeatureValue(
-                'explorer-sentry-tracing',
-                0
-            );
-            if (sampleRate > Math.random()) {
-                return event;
-            } else {
-                return null;
-            }
-        },
-        beforeSend(event) {
-            try {
-                // Filter out any code from unknown sources:
-                if (
-                    !event.exception?.values?.[0].stacktrace ||
-                    event.exception?.values?.[0].stacktrace?.frames?.[0]
-                        .filename === '<anonymous>'
-                ) {
-                    return null;
-                }
-            } catch (e) {}
-
-            return event;
-        },
-
-        denyUrls: [
-            // Chrome extensions
-            /extensions\//i,
-            /^chrome:\/\//i,
-            /<anonymous>/,
-        ],
-        allowUrls: [
-            /.*\.sui\.io/i,
-            /.*-mysten-labs\.vercel\.app/i,
-            'explorer-topaz.vercel.app',
-        ],
-    });
-}
-
-const adapters: WalletProviderProps['adapters'] = [
-    new WalletStandardAdapterProvider(),
-];
-if (import.meta.env.DEV) {
-    adapters.push(new UnsafeBurnerWalletAdapter());
-}
-
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
-        <WalletProvider adapters={adapters} autoConnect={false}>
-            <Router>
-                <App />
-            </Router>
-        </WalletProvider>
+        <RouterProvider router={router} />
     </React.StrictMode>
 );
