--- apps/explorer/src/styles/bytecode.module.css
@@ -1,3 +1,0 @@
-div.code {
-    @apply overflow-auto whitespace-pre border-solid border-gray-45 p-2 text-sm;
-}
