+++ apps/wallet/package.json
@@ -135,7 +135,7 @@
         "react-intl": "^6.0.5",
         "react-number-format": "^4.9.3",
         "react-redux": "^8.0.2",
-        "react-router-dom": "^6.3.0",
+        "react-router-dom": "^6.5.0",
         "react-textarea-autosize": "^8.3.4",
         "rxjs": "^7.5.6",
         "semver": "^7.3.8",
